utils/requirements.txt
faiss_cpu==1.11.0
gTTS==2.5.4
initialize==1.1.1
langchain_ollama==0.3.3
langdetect==1.0.9
numpy==2.3.1
pdfplumber==0.11.7
sentence_transformers==4.1.0
speechrecognition==3.14.3
wikipedia==1.4.0
